+++ CHANGELOG.md
@@ -8,6 +8,7 @@ backwards compatibility.
 
 * Fix equilibrium of wavetables with depth of 16 bits.
 * Visualize selected chord on module's display.
+* Visualize selected scale root on module's display.
 
 ## 0.6.0